opens up an image in the modal even via just the url
