+++ src/components/imageModal/index.css
@@ -1,3 +1,14 @@
 .modal-image {
   width: 100%;
 }
+
+.grayscale {
+  filter: gray;
+  -webkit-filter: grayscale(1);
+  filter: grayscale(1);
+}
+
+.grayscale: hover {
+  -webkit-filter: grayscale(0);
+  filter: none;
+ }
