+++ src/components/imageModal/index.js
@@ -1,14 +1,41 @@
-import React, { useState } from 'react';
+import React, { useEffect } from 'react';
 import { Modal, Button } from 'react-bootstrap';
+import { useSelector, useDispatch } from 'react-redux';
+import { closeModal, openModal } from "../../actions/imageModal";
+import { useParams } from 'react-router-dom'
 import './index.css'
 
 function ImageModal() {
-  const [modalShow, setModalShow] = useState(false)
+  const dispatch = useDispatch()
+  const modalShow = useSelector(state => state.imageModal.show)
+  const currentImage = useSelector(state => state.imageModal.currentImage)
+  const fetchError = useSelector(state => state.imageModal.error)
+
+  const params = useParams()
+  const imageId = params["id"]
+
+ useEffect(() => {
+    if (imageId) {
+      dispatch(openModal(imageId))
+    }
+  }, [dispatch, imageId])
+
+  function renderImageOrError() {
+    if (fetchError) {
+      return (
+        <><h1>{fetchError}</h1></>
+      )
+    } else {
+      return (
+        <><img className="modal-image" alt="" src={`${currentImage.url}/640/480`} /></>
+      )
+    }
+  }
 
   return (
     <Modal
       show={modalShow}
-      onHide={() => setModalShow(false)}
+      onHide={() => dispatch(closeModal())}
       size="lg"
       aria-labelledby="contained-modal-title-vcenter"
       centered
@@ -19,24 +46,13 @@ function ImageModal() {
         </Modal.Title>
       </Modal.Header>
       <Modal.Body>
-        <img className="modal-image" alt="" src="https://picsum.photos/id/634/640/480" />
+        {renderImageOrError()}
       </Modal.Body>
       <Modal.Footer>
-        <Button onClick={() => setModalShow(false)}>Close</Button>
+        <Button onClick={() => dispatch(closeModal())}>Close</Button>
       </Modal.Footer>
     </Modal>
   );
 }
 
 export default ImageModal;
-
-// .grayscale {
-//   filter: gray;
-//   -webkit - filter: grayscale(1);
-//   filter: grayscale(1)
-// }
-
-// .grayscale: hover {
-//   -webkit - filter: grayscale(0);
-//   filter: none;
-// }
